Añadido todos los headers

+++ frontend/javascript/captain-dashboard.js
@@ -5,15 +5,24 @@ document.addEventListener('DOMContentLoaded', () => {
         return;
     }
 
+    const userNameEl = document.getElementById('user-name');
+	
+    if (userNameEl) {
+        userNameEl.textContent = localStorage.getItem('nombre') || 'Capitán';
+    }
+
+    document.addEventListener('click', (e) => {
+        if(e.target.id === 'btn-edit'){
+            window.location.href = 'edit.html';
+            
+        } else if(e.target.id === 'btn-logout'){
+            localStorage.clear();
+            window.location.href = 'login.html';
+        }
+    });
+
     const nombre = localStorage.getItem('nombre') || 'Capitán';
-    const userNameEl    = document.getElementById('user-name');
     const welcomeNameEl = document.getElementById('welcome-name');
 
-    if (userNameEl)    userNameEl.textContent    = nombre;
     if (welcomeNameEl) welcomeNameEl.textContent = nombre;
-
-    document.getElementById('btn-logout').addEventListener('click', () => {
-        localStorage.clear();
-        window.location.href = 'login.html';
-    });
 });
